osd/retro: read input item state at its real width (big-endian fix)
The padX_get_state() callbacks cast item_internal to UINT8* and
dereferenced it, but the items they are registered against point
into the P1_state..P8_state arrays, which are UINT32[]. Reading a
UINT32 slot through a UINT8* returns the byte at the lowest
address: on little-endian hosts that is the LSB, so 0/1 button
values happened to come through intact, but on big-endian hosts
(e.g. Wii U / PPC) it is the always-zero MSB, so in-game input
read as permanently released while the RetroArch menu (which does
not go through these callbacks) still worked.

Match each callback's pointer type to its backing store: the pad
callbacks now read UINT32, and retrokbd_get_state now reads UINT16
(retrokbd_state[] is UINT16[]), which had the same latent bug.

Fixes #169. Reported and pad-side diagnosed by Ploggy.

+++ src/osd/retro/retromain.c
@@ -990,55 +990,55 @@ void init_input_descriptors(void)
 
 static INT32 pad1_get_state(void *device_internal, void *item_internal)
 {
-   UINT8 *itemdata = (UINT8 *)item_internal;
+   UINT32 *itemdata = (UINT32 *)item_internal;
    return *itemdata;
 }
 
 static INT32 pad2_get_state(void *device_internal, void *item_internal)
 {
-   UINT8 *itemdata = (UINT8 *)item_internal;
+   UINT32 *itemdata = (UINT32 *)item_internal;
    return *itemdata;
 }
 
 static INT32 pad3_get_state(void *device_internal, void *item_internal)
 {
-   UINT8 *itemdata = (UINT8 *)item_internal;
+   UINT32 *itemdata = (UINT32 *)item_internal;
    return *itemdata;
 }
 
 static INT32 pad4_get_state(void *device_internal, void *item_internal)
 {
-   UINT8 *itemdata = (UINT8 *)item_internal;
+   UINT32 *itemdata = (UINT32 *)item_internal;
    return *itemdata;
 }
 
 static INT32 pad5_get_state(void *device_internal, void *item_internal)
 {
-   UINT8 *itemdata = (UINT8 *)item_internal;
+   UINT32 *itemdata = (UINT32 *)item_internal;
    return *itemdata;
 }
 
 static INT32 pad6_get_state(void *device_internal, void *item_internal)
 {
-   UINT8 *itemdata = (UINT8 *)item_internal;
+   UINT32 *itemdata = (UINT32 *)item_internal;
    return *itemdata;
 }
 
 static INT32 pad7_get_state(void *device_internal, void *item_internal)
 {
-   UINT8 *itemdata = (UINT8 *)item_internal;
+   UINT32 *itemdata = (UINT32 *)item_internal;
    return *itemdata;
 }
 
 static INT32 pad8_get_state(void *device_internal, void *item_internal)
 {
-   UINT8 *itemdata = (UINT8 *)item_internal;
+   UINT32 *itemdata = (UINT32 *)item_internal;
    return *itemdata;
 }
 
 static INT32 retrokbd_get_state(void *device_internal, void *item_internal)
 {
-   UINT8 *itemdata = (UINT8 *)item_internal;
+   UINT16 *itemdata = (UINT16 *)item_internal;
    return *itemdata;
 }
